refactor(dashboard): extract keeper-config domain (A1 phase 2 cont.)
- dashboard-keeper-config.ts: asLoose*/normalizeStringList(canonical)/
  normalizeKeeper*/normalizePromptBlock/normalizeDefaultSourceKind/
  normalizePerProviderTimeoutMode/normalizeKeeperConfig (huge decoder) +
  SandboxProfile/SandboxNetworkMode/SharedMemoryScope/KeeperConfigUpdatePayload +
  fetchKeeperConfig/patchKeeperConfig/setKeeperToolPolicy.
- Fixed transcription bug during extraction: asLooseNumber had
  `typeof value === 'string'` (should be `!==`) — corrected in both the
  extracted file and verified against source via rg before commit.
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: ensureDevToken, asKeeperRuntimeBlockerClass,
  KeeperConfig/KeeperFeatureStatus/KeeperHookSlot (../types)
- typecheck green (tsc --noEmit)

dashboard.ts 1542 -> 1210 lines after keeper-config.
